requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
alembic==1.7.4
    # via flask-migrate
atomicwrites==1.4.0
    # via pytest
attrs==21.2.0
    # via pytest
backcall==0.2.0
    # via ipython
backports.entry-points-selectable==1.1.0
    # via virtualenv
black==21.9b0
    # via -r requirements.in
cfgv==3.3.1
    # via pre-commit
click==8.0.3
    # via
    #   black
    #   flask
    #   pip-tools
colorama==0.4.4
    # via
    #   click
    #   ipython
    #   pytest
coverage==6.0.2
    # via -r requirements.in
decorator==5.1.0
    # via ipython
distlib==0.3.3
    # via virtualenv
factory-boy==3.2.0
    # via
    #   -r requirements.in
    #   pytest-factoryboy
faker==9.5.1
    # via
    #   -r requirements.in
    #   factory-boy
    #   mixer
filelock==3.3.1
    # via virtualenv
flake8==4.0.1
    # via -r requirements.in
flask==2.0.2
    # via
    #   -r requirements.in
    #   flask-migrate
    #   flask-sqlalchemy
flask-migrate==3.1.0
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
greenlet==1.1.2
    # via sqlalchemy
identify==2.3.0
    # via pre-commit
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==1.1.1
    # via pytest
ipython==7.28.0
    # via -r requirements.in
isort==5.9.3
    # via -r requirements.in
itsdangerous==2.0.1
    # via flask
jedi==0.18.0
    # via ipython
jinja2==3.0.2
    # via flask
mako==1.1.5
    # via alembic
markupsafe==2.0.1
    # via
    #   jinja2
    #   mako
matplotlib-inline==0.1.3
    # via ipython
mccabe==0.6.1
    # via flake8
mixer==7.2.0
    # via -r requirements.in
mock==4.0.3
    # via -r requirements.in
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
packaging==21.0
    # via pytest
parso==0.8.2
    # via jedi
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pickleshare==0.7.5
    # via ipython
pip-tools==6.4.0
    # via -r requirements.in
platformdirs==2.4.0
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.15.0
    # via -r requirements.in
prompt-toolkit==3.0.20
    # via ipython
py==1.10.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pygments==2.10.0
    # via ipython
pyparsing==2.4.7
    # via packaging
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-env
    #   pytest-factoryboy
pytest-env==0.6.2
    # via -r requirements.in
pytest-factoryboy==2.1.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
pyyaml==6.0
    # via pre-commit
regex==2021.10.21
    # via black
six==1.16.0
    # via
    #   python-dateutil
    #   virtualenv
sqlalchemy==1.4.26
    # via
    #   alembic
    #   flask-sqlalchemy
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   pre-commit
    #   pytest
tomli==1.2.1
    # via
    #   black
    #   pep517
traitlets==5.1.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==3.10.0.2
    # via
    #   black
    #   faker
virtualenv==20.8.1
    # via pre-commit
wcwidth==0.2.5
    # via prompt-toolkit
werkzeug==2.0.2
    # via flask
wheel==0.37.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
